Move prompts, parsers, and formatters into process.py
